rename bluetooth files
